pci: pcie-brcmstb: Add basic support for BCM2712 PCIe
Add support for BCM2712 (Raspberry Pi 5) by adapting the Raspberry Pi's
dowstream kernel driver. Apart from resets handled by dedicated reset
controllers, there are couple of hacks that were needed to be ported
over from the reference implementation, most notably the function
adjusting the controller's reference clock and the BAR2 config address
remapping. None of these seem to be present in the upstream kernel
driver so eventually it'd be worthwile to check what does
U-Boot/downstream do differently that it's not needed in upstream (and
check that upstream really works with BCM2712 - it's still not merged
yet [1]).

Methods for handling the resets and address of the relocated hard debug
register are defined in new driver data structures that are implemented
in a similar fashion as in upstream/downstream Linux drivers.

[1] https://lore.kernel.org/lkml/20240910151845.17308-1-svarbanov@suse.de/

Signed-off-by: Jan Čermák <sairon@sairon.cz>